+++ src/main/scala/model/Passenger.scala
@@ -3,13 +3,12 @@ import java.time._
 import slick.jdbc.PostgresProfile.api._
 import scala.concurrent.Future
 
-case class Passenger (idPsg: Option[Long], name: String)
+case class Passenger (idPsg: Option[Int], name: String)
 
 class PassengerTable(tag: Tag) extends Table[Passenger](tag, "passenger") {
-  val idPsg = column[Long]("id_psg", O.PrimaryKey)
+  val idPsg = column[Int]("id_psg", O.PrimaryKey)
   val name = column[String]("name")
 
-  //val directorFk = foreignKey("director_id_fk", directorId, TableQuery[StaffTable])(_.id)
   def * = (idPsg.?, name) <> (Passenger.apply _ tupled, Passenger.unapply)
 }
