+++ src/main/scala/model/Trip.scala
@@ -5,11 +5,11 @@ import slick.jdbc.PostgresProfile.api._
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext.Implicits.global
 
-case class Trip (tripNo: Option[Long], idComp: Long, plane: String, townFrom: String, townTo: String, timeOut: LocalDateTime, timeIn: LocalDateTime)
+case class Trip (tripNo: Option[Int], idComp: Int, plane: String, townFrom: String, townTo: String, timeOut: LocalDateTime, timeIn: LocalDateTime)
 
 class TripTable(tag: Tag) extends Table[Trip](tag, "trip") {
-  val tripNo = column[Long]("trip_no", O.PrimaryKey)
-  val idComp = column[Long]("id_comp")
+  val tripNo = column[Int]("trip_no", O.PrimaryKey)
+  val idComp = column[Int]("id_comp")
   val plane = column[String]("plane")
   val townFrom = column[String]("town_from")
   val townTo = column[String]("town_to")
@@ -24,14 +24,13 @@ object TripTable {
   val table = TableQuery[TripTable]
 }
 
-case class TripToCompany (idComp: Option[Long], name: String)
+case class TripToCompany (idComp: Option[Int], name: String)
 class TripToCompanyTable(tag: Tag) extends Table[(TripToCompany)](tag, "trip_to_company") {
-  val idComp = column[Long]("id_comp", O.PrimaryKey)
+  val idComp = column[Int]("id_comp", O.PrimaryKey)
   val name = column[String]("name")
 
   val idCompFk = foreignKey("id_comp_fk", idComp, TableQuery[CompanyTable])(_.idComp)
   def * = (idComp.?, name) <> (TripToCompany.apply _ tupled,TripToCompany.unapply)
-  // ers - 1 parameter for previous tuple
 }
 
 object TripToCompanyTable {
